# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Integration Reliability Sandbox API
  version: 0.12.0
  summary: Submit synthetic events and watch them being delivered, retried, deduplicated and replayed.
  description: |
    A learning sandbox. Every request uses synthetic data, and the only delivery destination is a mock
    receiver inside the same app.

    **Asynchronous delivery.** `POST /v1/events` answers `202 Accepted` once the event and its delivery
    job are stored. Delivery happens later in a background worker. Poll the `statusUrl`
    (`GET /v1/events/{eventId}/deliveries`) until the latest delivery's `state` is `delivered` or `failed`.
    Nothing in this API waits for delivery to finish.

    **Retries.** Attempt 1 starts as soon as a worker picks the job up. Timeouts, connection failures,
    408, 429 and 5xx are retried after 2 s, 4 s and 8 s (deployment defaults), up to 4 attempts in total.
    Any other 4xx fails at once (`failureReason: non_retryable`). After the 4th failed attempt the delivery is
    `failed` / `attempts_exhausted`.

    **Errors.** Every error body has the same shape: `{"error": "<code>", "message": "<text>"}`, plus
    extra fields for some codes (`details`, `limit`, `state`, `replayDeliveryId`, `allow`, `path`). Branch
    on `error`; `message` is for people and may change.

    **Wrong method.** Any path below called with a method it doesn't list returns `405 method_not_allowed`
...

...
    in memory, per instance; `429 rate_limited` with `Retry-After`).

...

...
servers:
...
paths:
...
  /v1/sessions:
    post:
      tags: [Sessions]
      operationId: createSession
      summary: Create an anonymous demo session
      description: |
        Returns a bearer token **once**. Only its SHA-256 hash is stored, so it can't be shown again. It
        expires after 24 hours (deployment default). No request body.

        Limited to 10 creations per client IP per hour (in memory, per instance) and 1000 unexpired
        sessions in total (database).
      security: []
      responses:
        '201':
          description: Session created. Store `token` somewhere private, such as a secret environment variable.
          headers:
            Cache-Control: { $ref: '#/components/headers/NoStore' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NewSession' }
              example:
                token: irs_EXAMPLE_ONLY_not_a_real_token
                tokenType: Bearer
                createdAt: '2026-10-05T12:00:00.000Z'
                expiresAt: '2026-10-06T12:00:00.000Z'
                notice: Anonymous demo credential, shown only once. It is not a user account and expires automatically.
        '429': { $ref: '#/components/responses/RateLimited' }
        '503':
          description: Too many active demo sessions (`session_capacity_reached`).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: session_capacity_reached, message: The sandbox has too many active demo sessions. Try again later. }
  /v1/session:
    get:
      tags: [Sessions]
      operationId: getSession
      summary: Check your token and see when it expires
      responses:
        '200':
          description: The token is valid.
          content:
            application/json:
              schema:
                type: object
                required: [createdAt, expiresAt]
                properties:
                  createdAt: { type: string, format: date-time }
                  expiresAt: { type: string, format: date-time }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
components:
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
      description: |
        The demo token from `POST /v1/sessions` (`irs_` + 43 characters). A missing or non-Bearer header gets
        `401 missing_token`; an unknown, malformed **or expired** token gets `401 invalid_token`. Expired and
        invalid tokens are deliberately indistinguishable.
    receiverSecret:
      type: http
      scheme: bearer
      description: Internal. `RECEIVER_SECRET` from the server environment. Never given to clients.
    opsToken:
      type: http
      scheme: bearer
      description: Internal. `OPS_TOKEN` from the server environment (Render dashboard). Never given to clients.
...
  headers:
...
    NoStore:
      schema: { type: string, const: no-store }
  responses:
    Unauthorized:
      description: '`missing_token` (no `Authorization: Bearer` header) or `invalid_token` (unknown, malformed or expired token).'
      headers:
        WWW-Authenticate:
          schema: { type: string }
          example: Bearer realm="sandbox", error="invalid_token"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: invalid_token, message: The token is invalid or expired. Create a new demo session. }
...
    RateLimited:
      description: '`rate_limited`: the per-IP request budget is used up (in memory, per instance).'
      headers:
        Retry-After:
          description: Seconds until the window resets.
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: rate_limited, message: Too many requests. Try again in 42 seconds. }
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Stable machine-readable code. Branch on this.
          enum:
            - not_found
            - method_not_allowed
            - invalid_json
            - payload_too_large
            - bad_request
            - internal_error
            - rate_limited
            - session_capacity_reached
            - missing_token
            - invalid_token
            - idempotency_key_required
            - invalid_idempotency_key
            - unsupported_media_type
            - validation_failed
            - idempotency_key_conflict
            - event_limit_reached
            - invalid_query
            - delivery_not_failed
            - already_replayed
            - replay_limit_reached
            - receiver_unauthorized
            - unknown_session
            - simulated_server_error
            - simulated_slow_response
            - ops_unauthorized
        message: { type: string, description: For people. May change. }
        details:
          type: array
          items:
            type: object
            required: [field, issue]
            properties:
              field: { type: string, examples: [payload.title, query.limit, (body)] }
              issue: { type: string }
        limit: { type: integer }
        state: { $ref: '#/components/schemas/DeliveryState' }
        replayDeliveryId: { type: string, format: uuid }
        allow: { type: array, items: { type: string } }
        path: { type: string }
...
    NewSession:
      type: object
      required: [token, tokenType, createdAt, expiresAt, notice]
      properties:
        token:
          type: string
          description: 'Shown once. Keep it private. Send it as `Authorization: Bearer <token>`.'
        tokenType: { const: Bearer }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        notice: { type: string }
...
    DeliveryState:
      type: string
      enum: [pending, retry_scheduled, in_progress, delivered, failed]
      description: |
        - `pending`: waiting for its first attempt (or re-queued after a worker lost its lease).
        - `retry_scheduled`: a retryable attempt failed; the next one is due at `nextAttemptAt`.
        - `in_progress`: a worker is sending it now.
        - `delivered`: the receiver answered 2xx. Final.
        - `failed`: final; see `failureReason`. Can be replayed.
...
security:
  - sessionToken: []